# openapi/v1.yaml
components:
  responses:
    Forbidden:
      content: &id001
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
      description: Authenticated but not allowed (role, scope, tenant, quota).
    InternalError:
      content: *id001
      description: Unexpected error (a fixed body; details stay in the server log).
    Invalid:
      content: *id001
      description: The request is not valid.
    NotFound:
      content: *id001
      description: Not found (also for resources of another tenant).
    OtherError:
      content: *id001
      description: Any other error status, e.g. 409 conflict, 410 gone, 413 too large.
    TooManyRequests:
      content: *id001
      description: Rate limit or quota exceeded.
      headers:
        Retry-After:
          schema:
            minimum: 1
            type: integer
    Unauthorized:
      content: *id001
      description: Missing, invalid, expired or revoked credential.
      headers:
        WWW-Authenticate:
          schema:
            type: string
  schemas:
...
    Problem:
      additionalProperties: true
      description: RFC 9457 problem details. Never contains stack traces, SQL, host names or credentials
        (SEC-079).
      properties:
        detail:
          type: string
        errors:
          description: 'Validation errors (422): location and message, never the input.'
          items:
            properties:
              loc:
                items:
                  type: string
                type: array
              msg:
                type: string
            type: object
          type: array
        instance:
          description: The request path.
          type: string
        request_id:
          description: Also in the X-Request-Id header.
          type: string
        status:
          maximum: 599
          minimum: 400
          type: integer
        title:
          type: string
        type:
          description: urn:nf:problem:<kind> or about:blank
          type: string
      required:
      - type
      - title
      - status
      type: object
...
    ReportOut:
      properties:
        best:
          anyOf:
          - $ref: '#/components/schemas/BestOut'
          - type: 'null'
        cells:
          items:
            $ref: '#/components/schemas/CellOut'
          title: Cells
          type: array
        complete:
          title: Complete
          type: boolean
        factors:
          items:
            $ref: '#/components/schemas/FactorOut'
          title: Factors
          type: array
        metric:
          $ref: '#/components/schemas/MetricOut'
        pipeline:
          title: Pipeline
          type: string
        pipeline_version_id:
          title: Pipeline Version Id
          type: string
        prov_node_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Prov Node Id
        sensitivity:
          items:
            $ref: '#/components/schemas/SensitivityOut'
          title: Sensitivity
          type: array
        state:
          title: State
          type: string
        sweep_id:
          format: uuid
          title: Sweep Id
          type: string
        variants:
          items:
            $ref: '#/components/schemas/VariantSummary'
          title: Variants
          type: array
      required:
      - sweep_id
      - state
      - complete
      - pipeline
      - pipeline_version_id
      - metric
      - factors
      - cells
      - variants
      - sensitivity
      - best
      - prov_node_id
      title: ReportOut
      type: object
...
    RotatedOut:
      properties:
        previous_valid_until:
          description: Older secrets keep verifying until then (both signatures are sent).
          format: date-time
          title: Previous Valid Until
          type: string
        signing_secret:
          description: 'Shown once: the new signing secret.'
          title: Signing Secret
          type: string
        version:
          title: Version
          type: integer
      required:
      - signing_secret
      - version
      - previous_valid_until
      title: RotatedOut
      type: object
...
    SweepOut:
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        created_by:
          title: Created By
          type: string
        grid:
          additionalProperties:
            items: {}
            type: array
          title: Grid
          type: object
        id:
          format: uuid
          title: Id
          type: string
        metric:
          $ref: '#/components/schemas/MetricSpec'
        name:
          title: Name
          type: string
        pipeline:
          title: Pipeline
          type: string
        pipeline_version_id:
          title: Pipeline Version Id
          type: string
        prov_node_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Prov Node Id
        recording_ids:
          items:
            format: uuid
            type: string
          title: Recording Ids
          type: array
        run_ids:
          items:
            format: uuid
            type: string
          title: Run Ids
          type: array
        run_states:
          additionalProperties:
            type: integer
          title: Run States
          type: object
        state:
          description: queued | running | succeeded | partial | failed
          title: State
          type: string
        variants:
          items:
            $ref: '#/components/schemas/VariantOut'
          title: Variants
          type: array
      required:
      - id
      - name
      - state
      - run_states
      - pipeline
      - pipeline_version_id
      - recording_ids
      - grid
      - metric
      - variants
      - run_ids
      - prov_node_id
      - created_by
      - created_at
      title: SweepOut
      type: object
    SweepRequest:
      additionalProperties: false
      properties:
        grid:
          additionalProperties:
            items: {}
            type: array
          description: '{"<step>.<param>": [value, ...]}; the cartesian product gives the variants'
          maxProperties: 4
          minProperties: 1
          title: Grid
          type: object
        metric:
          $ref: '#/components/schemas/MetricSpec'
        name:
          anyOf:
          - maxLength: 200
            minLength: 1
            type: string
          - type: 'null'
          title: Name
        pipeline:
          description: 'base PipelineVersion: name@semver or pv ID'
          maxLength: 300
          minLength: 1
          title: Pipeline
          type: string
        recording_ids:
          items:
            format: uuid
            type: string
          maxItems: 32
          minItems: 1
          title: Recording Ids
          type: array
      required:
      - pipeline
      - recording_ids
      - grid
      - metric
      title: SweepRequest
      type: object
    SweepSummaryOut:
      description: One row of ``GET /v1/sweeps`` (open the sweep for states, variants and run IDs).
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        created_by:
          title: Created By
          type: string
        id:
          format: uuid
          title: Id
          type: string
        n_runs:
          title: N Runs
          type: integer
        n_variants:
          title: N Variants
          type: integer
        name:
          title: Name
          type: string
        pipeline_ref:
          title: Pipeline Ref
          type: string
        pipeline_version_id:
          title: Pipeline Version Id
          type: string
        prov_node_id:
          anyOf:
          - format: uuid
            type: string
          - type: 'null'
          title: Prov Node Id
      required:
      - id
      - name
      - pipeline_ref
      - pipeline_version_id
      - n_variants
      - n_runs
      - prov_node_id
      - created_by
      - created_at
      title: SweepSummaryOut
      type: object
...
    WebhookIn:
      properties:
        description:
          anyOf:
          - maxLength: 200
            type: string
          - type: 'null'
          title: Description
        event_types:
          items:
            const: run.finished
            type: string
          maxItems: 1
          minItems: 1
          title: Event Types
          type: array
        url:
          description: https URL with a public DNS name (private, loopback, link-local and metadata addresses
            are refused, SEC-076)
          maxLength: 2048
          minLength: 12
          title: Url
          type: string
      required:
      - url
      - event_types
      title: WebhookIn
      type: object
    WebhookIssued:
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        created_by:
          title: Created By
          type: string
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        disabled_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Disabled At
        event_types:
          items:
            type: string
          title: Event Types
          type: array
        id:
          format: uuid
          title: Id
          type: string
        key_versions:
          description: Signing keys currently valid.
          items:
            $ref: '#/components/schemas/KeyVersionOut'
          title: Key Versions
          type: array
        signing_secret:
          description: Shown once. HMAC-SHA-256 key for NF-Webhook-Signature (use its UTF-8 bytes).
          title: Signing Secret
          type: string
        url:
          title: Url
          type: string
      required:
      - id
      - url
      - description
      - event_types
      - created_by
      - created_at
      - disabled_at
      - key_versions
      - signing_secret
      title: WebhookIssued
      type: object
    WebhookOut:
      properties:
        created_at:
          format: date-time
          title: Created At
          type: string
        created_by:
          title: Created By
          type: string
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        disabled_at:
          anyOf:
          - format: date-time
            type: string
          - type: 'null'
          title: Disabled At
        event_types:
          items:
            type: string
          title: Event Types
          type: array
        id:
          format: uuid
          title: Id
          type: string
        key_versions:
          description: Signing keys currently valid.
          items:
            $ref: '#/components/schemas/KeyVersionOut'
          title: Key Versions
          type: array
        url:
          title: Url
          type: string
      required:
      - id
      - url
      - description
      - event_types
      - created_by
      - created_at
      - disabled_at
      - key_versions
      title: WebhookOut
      type: object
...
  securitySchemes:
    apiKey:
      bearerFormat: nfb_live_<public id>_<secret>
      description: Scoped, expiring API key (SEC-014). Revocation is effective at once for new requests
        and within 60 s for open streams (SEC-017).
      scheme: bearer
      type: http
    oidc:
      description: 'User access token from the identity provider (Authorization: Bearer).'
      openIdConnectUrl: https://id.example.invalid/.well-known/openid-configuration
      type: openIdConnect
info:
  description: 'Platform API v1. Data flows device -> SDK -> platform only: no operation, event or field
    sends anything to acquisition hardware (SEC-090). Errors are RFC 9457 application/problem+json. Only
    additive changes within v1; deprecations carry Deprecation and Sunset headers.'
  title: Platform API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /v1/sweeps:
    get:
      description: Sweeps of the caller's tenant, newest first (M3 open issue 5).
      operationId: listSweeps
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 200
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/SweepSummaryOut'
                title: Response Listsweeps
                type: array
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      - apiKey: []
      summary: List Sweeps
      x-nf-action: sweep:read
      x-nf-api-key-scope: data:read
    post:
      operationId: createSweep
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SweepRequest'
        required: true
      responses:
        '202':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SweepOut'
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      - apiKey: []
      summary: Create Sweep
      x-nf-action: sweep:create
      x-nf-api-key-scope: data:write
  /v1/sweeps/{sweep_id}:
    get:
      operationId: getSweep
      parameters:
      - in: path
        name: sweep_id
        required: true
        schema:
          format: uuid
          title: Sweep Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SweepOut'
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      - apiKey: []
      summary: Get Sweep
      x-nf-action: sweep:read
      x-nf-api-key-scope: data:read
  /v1/sweeps/{sweep_id}/report:
    get:
      operationId: getSweepReport
      parameters:
      - in: path
        name: sweep_id
        required: true
        schema:
          format: uuid
          title: Sweep Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportOut'
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      - apiKey: []
      summary: Get Sweep Report
      x-nf-action: sweep:read
      x-nf-api-key-scope: data:read
...
  /v1/webhooks:
    get:
      operationId: listWebhooks
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 50
          maximum: 200
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/WebhookOut'
                title: Response Listwebhooks
                type: array
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      summary: List Webhooks
      x-nf-action: webhook:read
    post:
      operationId: createWebhook
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WebhookIn'
        required: true
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookIssued'
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      summary: Create Webhook
      x-nf-action: webhook:create
  /v1/webhooks/{webhook_id}:
    delete:
      operationId: deleteWebhook
      parameters:
      - in: path
        name: webhook_id
        required: true
        schema:
          format: uuid
          title: Webhook Id
          type: string
      responses:
        '204':
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      summary: Delete Webhook
      x-nf-action: webhook:delete
    get:
      operationId: getWebhook
      parameters:
      - in: path
        name: webhook_id
        required: true
        schema:
          format: uuid
          title: Webhook Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookOut'
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      summary: Get Webhook
      x-nf-action: webhook:read
...
  /v1/webhooks/{webhook_id}/rotate-secret:
    post:
      operationId: rotateWebhookSecret
      parameters:
      - in: path
        name: webhook_id
        required: true
        schema:
          format: uuid
          title: Webhook Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RotatedOut'
          description: Successful Response
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Invalid'
        '429':
          $ref: '#/components/responses/TooManyRequests'
        '500':
          $ref: '#/components/responses/InternalError'
        default:
          $ref: '#/components/responses/OtherError'
      security:
      - oidc: []
      summary: Rotate Webhook Secret
      x-nf-action: webhook:update
...
servers:
- url: https://api.example.invalid
...